+++ back/src/main/java/com/back/global/security/CsrfCookieFilter.java
@@ -1,4 +1,4 @@
-package com.back.global.config;
+package com.back.global.security;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
